# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "De'Corrinna API",
        "description": "REST API backend De'Corrinna (Laravel 12 + PostgreSQL). Autentikasi memakai Bearer token (Laravel Sanctum).",
        "contact": {
            "name": "Tim Azatech",
            "email": "dev@azatech.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server lokal (dev)"
        }
    ],
    "paths": {
...
        "/api/admin/facilities": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Daftar fasilitas",
                "description": "Dikembalikan utuh tanpa paginasi — layar B4 menampilkannya sebagai deretan chip.",
                "operationId": "ff767b73ef1a2c51e15134f5e1a76a01",
                "responses": {
                    "200": {
                        "description": "Daftar fasilitas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Facility"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Tidak punya permission facilities:index",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Tambah fasilitas",
                "operationId": "447c6c50f3913e5f15754c07a436d3d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FacilityInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Fasilitas dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Facility"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/facilities/{facility}": {
            "put": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Ubah fasilitas",
                "operationId": "5938511b811707d2f63fac363e1cc526",
                "parameters": [
                    {
                        "name": "facility",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FacilityInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fasilitas diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Facility"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Hapus fasilitas",
                "description": "Soft delete. Tautan ke kategori ikut dilepas.",
                "operationId": "f2d5ca0c165eaf7f3d478c3414efe761",
                "parameters": [
                    {
                        "name": "facility",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Terhapus"
                    },
                    "404": {
                        "description": "Tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "title": "ErrorResponse",
                "description": "Envelope response error standar.",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Terjadi kesalahan."
                    },
                    "errors": {
                        "description": "Peta field -> daftar pesan (hanya untuk error validasi).",
                        "type": "object",
                        "example": {
                            "email": [
                                "Email wajib diisi."
                            ]
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Facility": {
                "title": "Facility",
                "description": "Cerminan FacilityResource.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Private Pool"
                    },
                    "icon": {
                        "description": "Nama ikon lucide.",
                        "type": "string",
                        "example": "waves",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "Aktif",
                        "enum": [
                            "Aktif",
                            "Nonaktif"
                        ]
                    }
                },
                "type": "object"
            },
            "FacilityInput": {
                "title": "FacilityInput",
                "description": "Body untuk POST/PUT fasilitas.",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Private Pool"
                    },
                    "icon": {
                        "type": "string",
                        "example": "waves",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "Aktif",
                        "enum": [
                            "Aktif",
                            "Nonaktif"
                        ]
                    },
                    "category_ids": {
                        "description": "Villa mana saja yang punya fasilitas ini.",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            },
...
            "SuccessResponse": {
                "title": "SuccessResponse",
                "description": "Envelope response sukses standar.",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token pribadi dari Laravel Sanctum. Kirim sebagai header: Authorization: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}